# Improve consistency in printing