Fixes #188

Before (Julia v1.4.2)
```julia
julia> N8f24[1/3 0.2]
1×2 Array{N8f24,2} with eltype Normed{UInt32,24}:
 0.333333  0.2

julia> N8f24[1/3, 0.2]
2-element Array{N8f24,1} with eltype Normed{UInt32,24}:
 0.33333333N8f24
 0.2N8f24

julia> FixedPoint[N0f8(1/3) N0f16(1/3)]
1×2 Array{Error showing value of type Array{FixedPoint,2}:
ERROR: MethodError: no method matching typechar(::Type{FixedPoint})

julia> FixedPoint[N0f8(1/3), N0f16(1/3)]
2-element Array{Error showing value of type Array{FixedPoint,1}:
ERROR: MethodError: no method matching typechar(::Type{FixedPoint})

julia> Normed{UInt128,28}(1)
1.0N100f28
```

After (Julia v1.4.2)
```julia
julia> N8f24[1/3 0.2]
1×2 Array{N8f24,2} with eltype Normed{UInt32,24}:
 0.333333  0.2

julia> N8f24[1/3, 0.2]
2-element Array{N8f24,1} with eltype Normed{UInt32,24}:
 0.33333333
 0.2

julia> FixedPoint[N0f8(1/3) N0f16(1/3)]
1×2 Array{FixedPoint,2}:
 0.333  0.33333

julia> FixedPoint[N0f8(1/3), N0f16(1/3)]
2-element Array{FixedPoint,1}:
 0.333N0f8
 0.33333N0f16

julia> Normed{UInt128,28}(1)
Normed{UInt128,28}(1.0)
```