@kimikage 2020-07-03
FYI
```julia
julia> Float32[1/3 0.2]
1×2 Array{Float32,2}:
 0.333333  0.2

julia> Float32[1/3, 0.2]
2-element Array{Float32,1}:
 0.33333334
 0.2

julia> AbstractFloat[Float16(1/3) Float32(1/3)]
1×2 Array{AbstractFloat,2}:
 0.3333  0.333333

julia> AbstractFloat[Float16(1/3), Float32(1/3)]
2-element Array{AbstractFloat,1}:
  Float16(0.3333)
 0.33333334f0

```