@kimikage 2020-07-04
`FixedPointNumbers.showtype()` had performance issues as well as functional issues.

Before
```julia
julia> @benchmark FixedPointNumbers.showtype(io, N0f8) # Julia v1.4.2
BenchmarkTools.Trial:
  memory estimate:  192 bytes
  allocs estimate:  4
  --------------
  minimum time:     266.134 ns (0.00% GC)
  median time:      276.998 ns (0.00% GC)
  mean time:        341.150 ns (2.18% GC)
  maximum time:     7.560 μs (92.47% GC)
  --------------
  samples:          10000
  evals/sample:     313

julia> @benchmark FixedPointNumbers.showtype(io, N0f8) # Julia v1.6.0-DEV
BenchmarkTools.Trial:
  memory estimate:  336 bytes
  allocs estimate:  6
  --------------
  minimum time:     441.414 ns (0.00% GC)
  median time:      466.667 ns (0.00% GC)
  mean time:        573.091 ns (11.50% GC)
  maximum time:     242.056 μs (99.73% GC)
  --------------
  samples:          10000
  evals/sample:     198
```

After
```julia
julia> @benchmark FixedPointNumbers.showtype(io, N0f8) # Julia v1.4.2
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     94.312 ns (0.00% GC)
  median time:      136.266 ns (0.00% GC)
  mean time:        130.076 ns (0.00% GC)
  maximum time:     33.543 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     932

julia> @benchmark FixedPointNumbers.showtype(io, N0f8) # Julia v1.6.0-DEV
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     139.721 ns (0.00% GC)
  median time:      142.276 ns (0.00% GC)
  mean time:        156.099 ns (0.00% GC)
  maximum time:     63.378 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     861
```

It's still slow but this PR (and upcoming changes in ColorTypes) can reduce the number of times where the type aliases are printed. Therefore, I will not optimize `showtype` with a string dictionary or something.