@kimikage 2020-07-21
I'll make sure that v0.8.4 doesn't have any fatal side effects on the downstream packages, then I'll merge this.